% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_periode.R
\name{calcul_periode}
\alias{calcul_periode}
\title{Calculer une statistique par période}
\usage{
calcul_periode(
  data,
  var = "col_calcul",
  group_vars = c("annee", "periode"),
  fun_name = c("mean", "sum", "min", "max"),
  nom = "col_resultat"
)
}
\arguments{
\item{data}{data.frame, tibble, \code{arrow::Dataset} ou \code{tbl_lazy}. Table contenant les données à agréger.}

\item{var}{Character. Nom de la colonne numérique sur laquelle appliquer la fonction d’agrégation.}

\item{group_vars}{Character vector. Noms des colonnes servant à regrouper les données (par exemple \code{c("annee", "periode")}).}

\item{fun_name}{Character. Nom de la fonction d’agrégation à appliquer. Doit être l’une des valeurs suivantes : \code{"mean"}, \code{"sum"}, \code{"min"}, \code{"max"}.}

\item{nom}{Character. Nom de la colonne résultat créée après agrégation.}
}
\value{
tibble avec les colonnes de regroupement et la colonne agrégée \code{nom}.
}
\description{
Cette fonction regroupe les données selon les variables spécifiées dans \code{group_vars} et applique une fonction d’agrégation sur une colonne donnée.
Compatible avec \code{data.frame} / \code{tibble}, \code{arrow::Dataset} et \code{duckdb} (\code{tbl_lazy})
Pour garantir la compatibilité avec Arrow et DuckDB, la fonction d’agrégation est fournie sous forme de chaîne de caractères (\code{fun_name}) et doit correspondre à une fonction d’agrégation standard supportée par ces moteurs.
}
\examples{


library(dplyr)

######################################################
# Exemple avec un data.table

data("ex_data_sim2")

resultat = ex_data_sim2 \%>\%
  def_periode(date_col = "DATE",
              periode = "saison") \%>\%
  calcul_periode(var = "T",
                 fun_name = "mean",
                 group_vars = c("LAMBX", "LAMBY", "annee", "periode"),
                 nom = "T_moy")

head(resultat)

######################################################
# Exemple avec un parquet / tbl_lazy


if (requireNamespace("duckdb", quietly = TRUE)) {
  
  library(DBI)
  
  # Chemin vers le parquet dans le package
  file <- system.file("ex_data_sim2.parquet", package = "franceclimat")
  
  # Connexion DuckDB en mémoire
  con <- dbConnect(duckdb::duckdb())
  
  # Création d'une table lazy directement depuis le parquet
  tbl_lazy <- tbl(con, sql(sprintf("SELECT * FROM read_parquet('\%s')", file)))
  
  resultat_tbl_lazy <- tbl_lazy \%>\%
    def_periode(date_col = "DATE", periode = "saison") \%>\%
    calcul_periode(
      var = "PRELIQ",
      fun_name = "sum",
      group_vars = c("LAMBX", "LAMBY", "annee", "periode"),
      nom = "PRELIQ_sum"
    ) \%>\%
    collect()
  
  dbDisconnect(con, shutdown = TRUE)
  
  head(resultat_tbl_lazy)
  
}

######################################################
# Exemple avec un fichier Parquet / Arrow

if (requireNamespace("arrow", quietly = TRUE)) {

  library(arrow)

  # Chemin vers le parquet dans le package
  file <- system.file("ex_data_sim2.parquet", package = "franceclimat")

  # Ouvrir le dataset
  ds <- arrow::open_dataset(file)

  # Ajouter les périodes (saison) puis collecter le résultat
  resultat_arrow <- ds \%>\%
    def_periode(date_col = "DATE", periode = "saison") \%>\%
    calcul_periode(
      var = "TSUP_H",
      fun = "mean",
      group_vars = c("LAMBX", "LAMBY", "annee", "periode"),
      nom = "TSUP_H_moy"
    ) \%>\%
    collect()

  head(resultat_arrow)

}

}
