% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcul_periode_multi.R
\name{calcul_periode_multi}
\alias{calcul_periode_multi}
\title{Calculer une statistique par période sur plusieurs cas}
\usage{
calcul_periode_multi(
  data,
  vars_funs,
  year_filter = c(1959, 2025),
  group_vars = c("LAMBX", "LAMBY", "annee", "periode")
)
}
\arguments{
\item{data}{data.frame, tibble, \code{arrow::Dataset} ou \code{tbl_lazy}. Table contenant les données.}

\item{vars_funs}{liste nommée, chaque élément correspond à une variable à agréger et doit contenir :
\describe{
\item{fun}{Caractère. Fonction d’agrégation à appliquer (\code{"mean"}, \code{"sum"}, \code{"min"}, \code{"max"}).}
\item{nom}{Caractère. Nom de la colonne résultat dans la table finale.}
}
Exemple : \code{list(T = list(fun = "mean", nom = "T_moy"),
                       PRELIQ = list(fun = "sum", nom = "PRELIQ_sum"))}}

\item{year_filter}{Numeric vector de longueur 2. Années minimales et maximales pour filtrer les données (par défaut \code{c(1959, 2025)}).}

\item{group_vars}{Character vector. Noms des colonnes servant à regrouper les données (par défaut \code{c("LAMBX", "LAMBY", "annee", "periode")}).}
}
\value{
Liste de tibbles. Chaque élément correspond à une table calculée pour une variable/fonction, avec les colonnes de regroupement et la colonne résultat renommée selon \code{nom}.
}
\description{
Cette fonction applique \code{calcul_periode()} sur plusieurs colonnes et fonctions spécifiées dans \code{vars_funs}. Elle retourne une liste de tables, chacune correspondant à une combinaison variable / fonction. Elle permet également de filtrer les données sur des années avec \code{year_filter}.
}
\examples{


######################################################
# Exemple avec un data.table

data("ex_data_sim2")

data = def_periode(
  data = ex_data_sim2,
  date_col = "DATE",
  periode = "annee"
)

vars_funs <- list(
  T = list(fun = "mean", nom = "T_moy"),
  PRELIQ = list(fun = "sum", nom = "PRELIQ_sum")
)

result_list <- calcul_periode_multi(data, vars_funs)

head(result_list)

######################################################
# Exemple avec un fichier Parquet / Arrow

if (requireNamespace("arrow", quietly = TRUE)) {
  
  # Chemin vers le parquet dans le package
  file <- system.file("ex_data_sim2.parquet", package = "franceclimat")
  
  # Ouvrir le dataset
  ds <- arrow::open_dataset(file)
  
  vars_funs <- list(
    TSUP_H = list(fun = "mean", nom = "Tmax_moy"),
    PRENEI = list(fun = "max", nom = "PRENEI_max")
  )
  
  # Ajouter les périodes
  ds <- def_periode(ds, date_col = "DATE", periode = "saison")
  
  # Appliquer calcul_periode_multi sur le dataset Arrow
  resultat_arrow <- calcul_periode_multi(
    data = ds,
    vars_funs = vars_funs,
    year_filter = c(1959, 2025),
    group_vars = c("LAMBX", "LAMBY", "annee", "periode")
  )
  
  head(resultat_arrow)

}


}
