% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_periode.R
\name{def_periode}
\alias{def_periode}
\title{Définir une période temporelle (année ou saison)}
\usage{
def_periode(data, date_col = "DATE", periode = "saison")
}
\arguments{
\item{data}{data.frame, tbl_lazy (DuckDB) ou Dataset (Dataset Arrow). Table contenant les données.}

\item{date_col}{Character. Nom de la colonne représentant la date (format AAAAMMJJ).}

\item{periode}{Character. Soit \code{"saison"} pour créer les saisons, soit \code{"annee"} pour créer simplement l'année.}
}
\value{
Même objet que \code{data} avec deux colonnes supplémentaires :
\describe{
\item{periode}{Période temporelle (Hiver, Printemps, Ete, Automne ou Annee)}
\item{annee}{Année associée à la période}
}
}
\description{
À partir d'une table contenant une colonne de dates au format AAAAMMJJ (année-mois-jour),
cette fonction ajoute deux colonnes :
\itemize{
\item \code{periode} : période temporelle ("Hiver", "Printemps", "Ete", "Automne" ou "Annee")
\item \code{annee} : année associée à la période
}
}
\details{
Lorsque \code{periode = "saison"}, les saisons météorologiques sont définies comme suit :
\itemize{
\item Hiver : décembre à février
\item Printemps : mars à mai
\item Été : juin à août
\item Automne : septembre à novembre
}

L'année associée à l'hiver correspond à l'année de janvier/février.  Ainsi, décembre 1960 est rattaché à l'hiver 1961.
}
\examples{


library(dplyr)

######################################################
# Exemple avec data.frame

data("ex_data_sim2")

resultat_df <- def_periode(
  data = ex_data_sim2,
  date_col = "DATE",
  periode = "saison"
)

######################################################
# Exemple avec DuckDB / parquet

if (requireNamespace("duckdb", quietly = TRUE)) {

  library(DBI)
  
  file <- system.file("ex_data_sim2.parquet", package = "franceclimat")

  con <- DBI::dbConnect(duckdb::duckdb())

  tbl_lazy <- dplyr::tbl(
    con,
    dbplyr::sql(sprintf("SELECT * FROM read_parquet('\%s')", file))
  )

  resultat_tbl_lazy <- tbl_lazy \%>\%
    def_periode("DATE", "saison") \%>\%
    collect()

  DBI::dbDisconnect(con, shutdown = TRUE)
}

######################################################
# Exemple avec Arrow

if (requireNamespace("arrow", quietly = TRUE)) {

  library(arrow)
  
  
  file <- system.file("ex_data_sim2.parquet", package = "franceclimat")

  ds <- arrow::open_dataset(file)

  resultat_arrow <- ds \%>\%
    def_periode("DATE", "saison") \%>\%
    collect()
}

}
