% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/franceclimat-package.R
\docType{package}
\name{franceclimat-package}
\alias{franceclimat}
\alias{franceclimat-package}
\title{franceclimat: Franceclimat}
\description{
Traitement et visualisation des données climatiques spatiales SIM produites par Météo France (modèle de simulation des schémas de surface Safran - Isba).
}
\author{
\strong{Maintainer}: Vincent Wawrzyniak \email{vincent.wawrzyniak@gmail.com}

}
\keyword{internal}
