% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_lineaire.R
\name{regression_lineaire}
\alias{regression_lineaire}
\title{Effectuer une régression linéaire par groupe}
\usage{
regression_lineaire(
  data,
  var = "col_calcul",
  group_vars = c("LAMBX", "LAMBY", "periode")
)
}
\arguments{
\item{data}{Un \code{data.frame} ou \code{tibble} contenant les colonnes à analyser.}

\item{var}{Character. Nom de la colonne numérique à utiliser comme variable dépendante dans la régression.}

\item{group_vars}{Character vector. Noms des colonnes servant à grouper les données avant la régression.}
}
\value{
Un tibble avec les colonnes de regroupement, la pente (\code{slope}) et la valeur-p (\code{p_value}) pour chaque groupe.
}
\description{
Cette fonction effectue une régression linéaire du type \code{var ~ annee} pour chaque combinaison des colonnes spécifiées dans \code{group_vars}. Elle retourne la pente (estimate) et la valeur-p associée à la variable temporelle \code{annee}.
}
\examples{


data("ex_data_sim2")

data = ex_data_sim2 \%>\%
  def_periode(date_col = "DATE",
              periode = "saison") \%>\%
  calcul_periode(var = "T",
                 fun_name = "mean",
                 group_vars = c("LAMBX", "LAMBY", "annee", "periode"),
                 nom = "T_moy")

res_lm <- regression_lineaire(
  data = data,
  var = "T_moy",
  group_vars = c("LAMBX", "LAMBY", "periode")
)


}
