% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_periode.R
\name{def_periode}
\alias{def_periode}
\title{Définir une période temporelle (année ou saison)}
\usage{
def_periode(data, date_col = "DATE", periode = "saison")
}
\arguments{
\item{data}{data.frame. Table avec les données qui doit contenir un colonne représentant la date au format année-mois-jour (AAAAMMJJ).}

\item{date_col}{Character. Nom de la colonne représentant la date.}

\item{periode}{Character. saison ou annee}
}
\value{
tibble. Table initiale avec 2 colonnes en plus : periode et annee
}
\description{
À partir d'une table contenant une colonne de dates au format #' AAAAMMJJ (année-mois-jour), cette fonction ajoute deux colonnes :
\itemize{
\item \code{periode} : période temporelle ("Hiver", "Printemps", "Été",
"Automne" ou "Année")
\item \code{annee} : année associée à la période
}
}
\details{
Lorsque \code{periode = "saison"}, les saisons météorologiques sont définies comme suit :
\itemize{
\item Hiver : décembre à février
\item Printemps : mars à mai
\item Été : juin à août
\item Automne : septembre à novembre
}

L'année associée à l'hiver correspond à l'année de janvier/février. Ainsi, décembre 1960 est rattaché à l'hiver 1961.
}
\examples{


data("ex_data_sim2")

resultat = def_periode(data = ex_data_sim2,
                       date_col = "DATE",
                       periode = "saison")
}
